Advanced wave form widget. Still some issues with redrawing.

+++ code/engine/viaudioconnections/viaudioconnection/headers/viaudioposition.h
@@ -6,7 +6,7 @@
 class ViAudioPosition
 {
 	public:
-		ViAudioPosition(qint64 bytes = 0, qint8 bitDepth = 0, qint16 secondsInByte = 0);
+		ViAudioPosition(qint64 bytes = 0, qint8 bitDepth = 0, qreal secondsInByte = 0);
 		qint64 seconds();
 		qint64 milliseconds();
 		qint64 bytes();
